+++ example/lib/app/app_widget.dart
@@ -7,15 +7,6 @@ class AppWidget extends StatelessWidget {
     return MaterialApp(
       initialRoute: "/",
       onGenerateRoute: Modular.generateRoute,
-      onUnknownRoute: (context) {
-        return MaterialPageRoute(
-          builder: (context) => Scaffold(
-            appBar: AppBar(
-              title: Text('Página não encontrada'),
-            ),
-          ),
-        );
-      },
     );
   }
 }