+++ example/lib/app/modules/home/home_module.dart
@@ -12,9 +12,10 @@ class HomeModule extends ChildModule {
 
   @override
   List<Router> get routers => [
-    Router("/", child: (_, args) => HomeWidget()),
+    Router("/:id/:id2", child: (_, args) => HomeWidget()),
     Router("/list", child: (_, args) => ListWidget()),
-    Router("/list/1", child: (_, args) => ListWidget(param: "Outro Widget",)),
+    Router("/again", module: HomeModule()),
+    Router("/list/:id", child: (_, args) => ListWidget(param: args.params['id'],)),
   ];
 
   static Inject get to => Inject<HomeModule>.of();