+++ lib/flutter_modular.dart
@@ -4,6 +4,7 @@ export 'src/modular_base.dart';
 export 'src/interfaces/child_module.dart';
 export 'src/interfaces/main_module.dart';
 export 'src/interfaces/disposable.dart';
+export 'src/interfaces/route_guard.dart';
 export 'src/inject/inject.dart';
 export 'src/inject/bind.dart';
 export 'src/routers/router.dart';