+++ lib/src/inject/bind.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter_modular/flutter_modular.dart';
 
 class Bind {