+++ lib/src/interfaces/main_module.dart
@@ -2,7 +2,6 @@ import 'package:flutter/widgets.dart';
 
 import 'child_module.dart';
 
-
 abstract class MainModule extends ChildModule {
   Widget get bootstrap;
 }