+++ test/app/app_module.dart
@@ -3,6 +3,8 @@ import 'package:flutter_modular/flutter_modular.dart';
 
 import 'app_bloc.dart';
 import 'app_widget.dart';
+import 'guard/guard.dart';
+import 'modules/forbidden/forbidden_widget.dart';
 import 'modules/home/home_module.dart';
 
 class AppModule extends MainModule {
@@ -13,7 +15,13 @@ class AppModule extends MainModule {
 
   @override
   List<Router> get routers => [
-        Router("/", module: HomeModule()),
+        Router(
+          "/forbidden",
+          child: (_, args) => ForbiddenWidget(),
+          guards: [MyGuard()],
+          transition: TransitionType.fadeIn,
+        ),
+        Router("/", module: HomeModule(), transition: TransitionType.fadeIn,),
         Router("/home", module: HomeModule()),
       ];