+++ test/app/modules/home/home_module.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_modular/flutter_modular.dart';
 
+import '../product/product_module.dart';
 import 'home_bloc.dart';
 import 'home_widget.dart';
 
@@ -11,8 +12,9 @@ class HomeModule extends ChildModule {
 
   @override
   List<Router> get routers => [
-    Router("/", child: (_, args) => HomeWidget()),
-    Router("/list", child: (_, args) => HomeWidget()),
+    Router("/", child: (_, args) => HomeWidget(), transition: TransitionType.fadeIn,),
+    Router("/list/:id/:id2", child: (_, args) => HomeWidget()),
+    Router("/product", module: ProductModule()),
   ];
 
   static Inject get to => Inject<HomeModule>.of();