+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [0.0.11] - 15 Dec 2019
+
+* Added Route Transitions.
+
+## [0.0.10] - 14 Dec 2019
+
+* Added logo
+
+## [0.0.8] - 13 Dec 2019
+
+* Route Guard
+
+## [0.0.7] - 10 Dec 2019
+
+* Dynamic Router
+* Added Doc Translation
+* Change BrowserModule to MainModule
+* Change CommonModule to ChildModule
+* Corrigido erro de blink na primeira rota
+* fix routes param
+
 ## [0.0.1] - 8 Dec 2019
 
 * First Release